HealthGuard - Demo

+++ HealthGuard/Models/Entity/DiagnosisResult.cs
@@ -6,14 +6,14 @@ namespace HealthGuard.Models.Entity
     public class DiagnosisResult
     {
         [Key]
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         [ForeignKey("DiagnosticSession")]
-        public int SessionId { get; set; }
+        public long SessionId { get; set; }
         public virtual DiagnosticSession DiagnosticSession { get; set; }
 
         [ForeignKey("Disease")]
-        public int DiseaseId { get; set; }
+        public long DiseaseId { get; set; }
         public virtual Disease Disease { get; set; }
 
         public double ProbabilityPercentage { get; set; } // VD: 85.5%
